Add files via upload
The risk management approach's goal is to assess LEGION Broadband Network Company's flaws and vulnerabilities. As well as recognizing potential risks and their expected fiscal impact on the organization, as well as ensuring the security of linked systems. The study is organized and presents both quantitative and qualitative risk assessment methodologies.